Move update one-many tests to new dir
No test changes, all were already using the new framework
